requirements.txt
Flask
Pillow
Werkzeug